// src/types.ts
export interface CompanyColors {
    id: string;
    title: string | null;
    accent: string | null;
    form_bg: string | null;
    surface: string | null;
    text_dim: string | null;
    background: string | null;
    company_id: string;
    text_color: string | null;
    button_text: string | null;
    primary_color: string | null;
    secondary_color: string | null;
    company_name?: string;
}